requirements.txt
Flask==2.0.1
firebase-admin==6.1.0
python-dotenv==0.19.0
boltiotai
